Update existing `attackby()` calls

+++ code/game/objects/effects/chem/foam.dm
@@ -161,20 +161,26 @@
 		to_chat(user, "<span class='notice'>You hit the metal foam but bounce off it.</span>")
 	return
 
-/obj/structure/foamedmetal/attackby(var/obj/item/I, var/mob/user)
-	if(istype(I, /obj/item/grab))
-		var/obj/item/grab/G = I
-		G.affecting.loc = src.loc
-		visible_message("<span class='warning'>[G.assailant] smashes [G.affecting] through the foamed metal wall.</span>")
-		qdel(I)
-		qdel(src)
-		return
 
-	if(prob(I.force * 20 - metal * 25))
-		user.visible_message("<span class='warning'>[user] smashes through the foamed metal.</span>", "<span class='notice'>You smash through the foamed metal with \the [I].</span>")
-		qdel(src)
-	else
-		to_chat(user, "<span class='notice'>You hit the metal foam to no effect.</span>")
+/obj/structure/foamedmetal/grab_attack(obj/item/grab/G)
+	G.affecting.loc = loc
+	visible_message("<span class='warning'>[G.assailant] smashes [G.affecting] through the foamed metal wall.</span>")
+	qdel(G)
+	qdel(src)
+	return TRUE
+
+
+/obj/structure/foamedmetal/use_weapon(obj/item/I, mob/user, click_params)
+	if (!(I.item_flags & ITEM_FLAG_NO_BLUDGEON))
+		if(prob(I.force * 20 - metal * 25))
+			user.visible_message("<span class='warning'>[user] smashes through the foamed metal.</span>", "<span class='notice'>You smash through the foamed metal with \the [I].</span>")
+			qdel(src)
+			return TRUE
+		else
+			to_chat(user, "<span class='notice'>You hit the metal foam to no effect.</span>")
+			return FALSE
+	return ..()
+
 
 /obj/structure/foamedmetal/CanPass(atom/movable/mover, turf/target, height=1.5, air_group = 0)
 	if(air_group)
